+++ cmd/oidc-webhook-authenticator/app/oidc_webhook_authenticator.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package authentication
+package app
 
 import (
 	"context"
@@ -29,6 +29,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 
 	authenticationv1alpha1 "github.com/gardener/oidc-webhook-authenticator/apis/authentication/v1alpha1"
+	authcontroller "github.com/gardener/oidc-webhook-authenticator/controllers/authentication"
 	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -114,7 +115,7 @@ func run(ctx context.Context, opts *options.Config, setupLog logr.Logger) error 
 		return err
 	}
 
-	authReconciler := &OpenIDConnectReconciler{
+	authReconciler := &authcontroller.OpenIDConnectReconciler{
 		Client: mgr.GetClient(),
 		Log:    ctrl.Log.WithName("controllers").WithName("OpenIDConnect"),
 		Scheme: mgr.GetScheme(),
